fix: Avoid import from app in server
