+++ app/components/Sharing/PublicAccess.tsx
@@ -8,6 +8,7 @@ import { Trans, useTranslation } from "react-i18next";
 import { Link } from "react-router-dom";
 import { toast } from "sonner";
 import styled, { useTheme } from "styled-components";
+import Squircle from "@shared/components/Squircle";
 import { s } from "@shared/styles";
 import { UrlHelper } from "@shared/utils/UrlHelper";
 import Document from "~/models/Document";
@@ -21,7 +22,6 @@ import { AvatarSize } from "../Avatar/Avatar";
 import CopyToClipboard from "../CopyToClipboard";
 import NudeButton from "../NudeButton";
 import { ResizingHeightContainer } from "../ResizingHeightContainer";
-import Squircle from "../Squircle";
 import Text from "../Text";
 import Tooltip from "../Tooltip";
 import { StyledListItem } from "./MemberListItem";
