+++ shared/i18n/locales/en_US/translation.json
@@ -238,7 +238,7 @@
   "{{authorName}} created <3></3>": "{{authorName}} created <3></3>",
   "{{authorName}} opened <3></3>": "{{authorName}} opened <3></3>",
   "Show menu": "Show menu",
-  "Choose icon": "Choose icon",
+  "Choose an icon": "Choose an icon",
   "Filter": "Filter",
   "Loading": "Loading",
   "Select a color": "Select a color",
